+++ src/linear_algebra/COOSymmetricMatrix.hpp
@@ -10,10 +10,10 @@ public:
    std::vector<int> row_indices;
    std::vector<int> column_indices;
 
-   COOSymmetricMatrix(size_t dimension, size_t capacity);
+   COOSymmetricMatrix(int dimension, size_t capacity);
 
-   void for_each(const std::function<void (size_t, size_t, double)>& f) const override;
-   void insert(double term, size_t row_index, size_t column_index) override;
+   void for_each(const std::function<void (int, int, double)>& f) const override;
+   void insert(double term, int row_index, int column_index) override;
 
    double norm_1();
    /*COOSymmetricMatrix add_identity_multiple(double multiple);*/
